+++ hammer/editgameclass.h
@@ -79,8 +79,8 @@ class CEditGameClass
 		// Interface to entity connections.
 		//
 		void Connections_Add(CEntityConnection *pConnection);
-		inline int Connections_GetCount(void);
-		inline CEntityConnection *Connections_Get(int nIndex);
+		inline intp Connections_GetCount(void);
+		inline CEntityConnection *Connections_Get(intp nIndex);
 		bool Connections_Remove(CEntityConnection *pConnection);
 		void Connections_RemoveAll(void);
 		void Connections_FixBad(bool bRelink = true);
@@ -89,8 +89,8 @@ class CEditGameClass
 		// Interface to entity connections.
 		//
 		void Upstream_Add(CEntityConnection *pConnection);
-		inline int Upstream_GetCount(void);
-		inline CEntityConnection *Upstream_Get(int nIndex);
+		inline intp Upstream_GetCount(void);
+		inline CEntityConnection *Upstream_Get(intp nIndex);
 		bool Upstream_Remove(CEntityConnection *pConnection);
 		void Upstream_RemoveAll(void);
 		void Upstream_FixBad();
@@ -139,7 +139,7 @@ class CEditGameClass
 //-----------------------------------------------------------------------------
 // Purpose: Returns the number of input/output connections that this object has.
 //-----------------------------------------------------------------------------
-int CEditGameClass::Connections_GetCount(void)
+intp CEditGameClass::Connections_GetCount(void)
 {
 	return m_Connections.Count();
 }
@@ -148,7 +148,7 @@ int CEditGameClass::Connections_GetCount(void)
 //-----------------------------------------------------------------------------
 // Purpose: Returns the number of input/output connections that this object has.
 //-----------------------------------------------------------------------------
-CEntityConnection *CEditGameClass::Connections_Get(int nIndex)
+CEntityConnection *CEditGameClass::Connections_Get(intp nIndex)
 {
 	return m_Connections.Element(nIndex);
 }
@@ -157,7 +157,7 @@ CEntityConnection *CEditGameClass::Connections_Get(int nIndex)
 //-----------------------------------------------------------------------------
 // Purpose: Returns the number of input/output connections that this object has.
 //-----------------------------------------------------------------------------
-int CEditGameClass::Upstream_GetCount(void)
+intp CEditGameClass::Upstream_GetCount(void)
 {
 	return m_Upstream.Count();
 }
@@ -166,7 +166,7 @@ int CEditGameClass::Upstream_GetCount(void)
 //-----------------------------------------------------------------------------
 // Purpose: Returns the number of input/output connections that this object has.
 //-----------------------------------------------------------------------------
-CEntityConnection *CEditGameClass::Upstream_Get(int nIndex)
+CEntityConnection *CEditGameClass::Upstream_Get(intp nIndex)
 {
 	return m_Upstream.Element(nIndex);
 }
